CV submodule

Documentation for SDtoolbox.jl

SDtoolbox.CV — Module

Shock and Detonation Toolbox "cv" module

Calculates constant-volume explosions.

This module defines the following functions:

cvsolve

################################################################################ Theory, numerical methods and applications are described in the following report:

SDToolbox Numerical Tools for Shock and Detonation Wave Modeling,
Explosion Dynamics Laboratory, Contributors:
S. Browne, J. Ziegler, N. Bitter, B. Schmidt, J. Lawson and J. E. Shepherd,
GALCIT Technical Report FM2018.001 Revised January 2021.
California Institute of Technology, Pasadena, CA USA

Please cite this report and the website if you use these routines.

Please refer to LICENCE.txt or the above report for copyright and disclaimers.

http://shepherd.caltech.edu/EDL/PublicResources/sdt/

SDtoolbox.CV.cv! — Method

cv!(dy::Vector{Float64},y::Vector{Float64},params,t::Real)::Vector{Float64}

Evaluates the system of ordinary differential equations for an adiabatic, constant-volume, zero-dimensional reactor. It assumes that the "gas" object represents a reacting ideal gas mixture.

INPUT: t = time y = solution array [temperature, species mass 1, 2, ...] gas = working gas object

OUTPUT: An array containing time derivatives of: temperature and species mass fractions, formatted in a way that the integrator in cvsolve can recognize.

SDtoolbox.CV.cvsolve — Method

cvsolve(gas::PyObject;t_end::Real=1e-6,max_step::Real=1e-5,
            t_eval=nothing,relTol::Real=1e-5,absTol::Real=1e-8)

Solves the ODE system defined in cv!, taking the gas object input as the initial state.

FUNCTION SYNTAX: output = cvsolve(gas,**kwargs)

INPUT: gas = working gas object

OPTIONAL INPUT: tend = end time for integration, in sec maxstep = maximum time step for integration, in sec t_eval = array of time values to evaluate the solution at. If left as "None", solver will select values. Sometimes these may be too sparse for good-looking plots. relTol = relative tolerance absTol = absolute tolerances

OUTPUT: output = a dictionary containing the following results: time = time array T = temperature profile array P = pressure profile array speciesY = species mass fraction array speciesX = species mole fraction array

    gas = working gas object

    exo_time = pulse width (in secs) of temperature gradient (using 1/2 max)
    ind_time = time to maximum temperature gradient
    ind_time_10 = time to 10% of maximum temperature gradient
    ind_time_90 = time to 90% of maximum temperature gradient