Postshock submodule

Documentation for SDtoolbox.jl

SDtoolbox.Postshock — Module

Shock and Detonation Toolbox "Postshock" module

Calculates CJ detonation speed and post-shock states for frozen and equilibrium cases.

This module defines the following functions:

LSQ_CJspeed
hug_fr
hug_eq
FHFP
CJ_calc
CJspeed
PostShock_fr
PostShock_eq
shock_calc
shk_eq_calc

################################################################################ Theory, numerical methods and applications are described in the following report:

SDToolbox Numerical Tools for Shock and Detonation Wave Modeling,
Explosion Dynamics Laboratory, Contributors:
S. Browne, J. Ziegler, N. Bitter, B. Schmidt, J. Lawson and J. E. Shepherd,
GALCIT Technical Report FM2018.001 Revised January 2021.
California Institute of Technology, Pasadena, CA USA

Please cite this report and the website if you use these routines.

Please refer to LICENCE.txt or the above report for copyright and disclaimers.

http://shepherd.caltech.edu/EDL/PublicResources/sdt/

SDtoolbox.Postshock.CJ_calc — Method

CJ_calc(gas, gas1, ERRFT, ERRFV, x)

Calculates the Chapman-Jouguet wave speed using Reynolds' iterative method.

FUNCTION SYNTAX: [gas,w1] = CJ_calc(gas,gas1,ERRFT,ERRFV,x)

INPUT: gas = working gas object gas1 = gas object at initial state ERRFT,ERRFV = error tolerances for iteration x = density ratio

OUTPUT: gas = gas object at equilibrium state w1 = initial velocity to yield prescribed density ratio

SDtoolbox.Postshock.CJspeed — Method

defaults to CEA Algorithm

SDtoolbox.Postshock.FHFP — Method

FHFP(w1,gas2,gas1)

Uses the momentum and energy conservation equations to calculate error in pressure and enthalpy given shock speed, upstream (gas1) and downstream states (gas2). States are not modified by these routines.

FUNCTION SYNTAX: [FH,FP] = FHFP(w1,gas2,gas1)

INPUT: w1 = shock speed (m/s) gas2 = gas object at working/downstream state gas1 = gas object at initial/upstream state

OUTPUT: FH,FP = error in enthalpy and pressure

SDtoolbox.Postshock.FHFP_CJ2 — Method

FHFP_CJ2(gas,gas1,gas2)

FHFP_CJ2 Uses the momentum and energy conservation equations and the equilibrium sound speed to calculate error in current pressure and enthalpy guesses. In this case, state 2 is in equilibrium.

FUNCTION SYNTAX [FH,FP,cjspeed] = FHFPCJ2(gas,gas1,gas2)

INPUT gas = working gas object gas1 = gas object at initial state gas2 = dummy gas object (for calculating numerical derivatives)

OUTPUT FH,FP = error in enthalpy and pressure cj_speed = CJ detonation speed (m/s)

SDtoolbox.Postshock.LSQ_CJspeed — Method

LSQ_CJspeed(x,y)

Determines least squares fit of parabola to input data

FUNCTION SYNTAX: [a,b,c,R2,SSE,SST] = LSQ_CJspeed(x,y)

INPUT: x = independent data points y = dependent data points

OUTPUT: a,b,c = coefficients of quadratic function (ax^2 + bx + c = 0) R2 = R-squared value SSE = sum of squares due to error SST = total sum of squares

SDtoolbox.Postshock.PostShock_eq — Method

PostShock_eq(U1, P1, T1, q, mech)

Calculates equilibrium post-shock state for a specified shock velocity and pre-shock state.

FUNCTION SYNTAX: gas = PostShock_eq(U1,P1,T1,q,mech)

INPUT: U1 = shock speed (m/s) P1 = initial pressure (Pa) T1 = initial temperature (K) q = reactant species mole fractions in one of Cantera's recognized formats mech = cti file containing mechanism data (e.g. 'gri30.cti')

OUTPUT: gas = gas object at equilibrium post-shock state

SDtoolbox.Postshock.PostShock_fr — Method

PostShock_fr(U1, P1, T1, q, mech)

Calculates frozen post-shock state for a specified shock velocity and pre-shock state.

FUNCTION SYNTAX: gas = PostShock_fr(U1,P1,T1,q,mech)

INPUT: U1 = shock speed (m/s) P1 = initial pressure (Pa) T1 = initial temperature (K) q = reactant species mole fractions in one of Cantera's recognized formats mech = cti file containing mechanism data (e.g. 'gri30.cti')

OUTPUT: gas = gas object at frozen post-shock state

SDtoolbox.Postshock._CJspeed_CEA — Method

_CJspeed_CEA(P1,T1,X1,mech)

This function calculates the CJ-detonation velocity for given initial conditions.

It is based on (i.e. mostly copied from) the CJ-detonation part of the NASA CEA Fortran95 code, available at http://www.grc.nasa.gov/WWW/CEAWeb/ceaHome.htm, which itself is based on "Calculation of Detonation Properties and Effect of Independent Parameters on Gaseous Detonations" by Frank J. Zeleznik and Sanford Gordon, ARS Journal, April 1962.

INPUT: P1 = pressure in (Pa) T1 = temperature in (K) X1 = mole fraction column vector mech = file name of kinetic mechanism, e.g. 'gri30.xml'

OUTPUT: cj_speed = CJ detonation velocity in (m/s)

SDtoolbox.Postshock._CJspeed_aeq — Method

_CJspeed_aeq(P1, T1, q, mech)

CJspeed_aeq Calculates CJ detonation velocity and CJ state based on equilibrium sound speed

FUNCTION SYNTAX cj_speed,gas = CJspeed(P1,T1,q,mech)

INPUT P1 = initial pressure (Pa) T1 = initial temperature (K) q = string of reactant species mole fractions mech = cti file containing mechanism data (i.e. 'gri30.cti')

OUTPUT cj_speed = CJ detonation speed (m/s)

SDtoolbox.Postshock._CJspeed_umin — Method

_CJspeed_umin(P1, T1, q, mech; fullOutput=false)

Calculates CJ detonation velocity for a given pressure, temperature, and composition.

FUNCTION SYNTAX: If only CJ speed required: cjspeed = CJspeed(P1,T1,q,mech) If full output required: [cjspeed,R2,plot_data] = CJspeed(P1,T1,q,mech,fullOutput=true)

INPUT: P1 = initial pressure (Pa) T1 = initial temperature (K) q = reactant species mole fractions in one of Cantera's recognized formats mech = cti file containing mechanism data (e.g. 'gri30.cti')

OPTIONAL INPUT: fullOutput = set true for R-squared value and pre-formatted plot data (the latter for use with sdtoolbox.utilities.CJspeed_plot)

OUTPUT cjspeed = CJ detonation speed (m/s) R2 = R-squared value of LSQ curve fit (optional) plotdata = tuple (rr,w1,dnew,a,b,c) rr = density ratio w1 = speed dnew = minimum density a,b,c = quadratic fit coefficients

SDtoolbox.Postshock.equilSoundSpeeds — Method

equilSoundSpeeds(gas)

Calculates equilibrium and frozen sound speeds. For the equilibrium sound speed, the gas is equilibrated holding entropy and specific volume constant.

FUNCTION SYNTAX [aequil,afrozen] = equilSoundSpeeds(gas)

INPUT gas = working gas object (modified inside function)

OUTPUT aequil = equilibrium sound speed (m/s) afrozen = frozen sound speed (m/s)

SDtoolbox.Postshock.hug_eq — Method

hug_eq(x,vb,h1,P1,v1,gas)

Computes difference in enthalpy computed from assumed (T, V) state and equilibrium composition state hugoniot evaluation. Used with root solver such as 'fsolve' to compute equilibrium hugoniot as a function of volume. Input gas object is modified to correspond to input (T, V) and an equilibrium composition.

FUNCTION SYNTAX: diff = hug_eq(x,vb,h1,P1,v1,gas)

USAGE: fval = fsolve(hug_eq,Ta,args=(vb,h1,P1,v1,gas)) = equilibrium Hugoniot temperature (K) corresponding to vb

INPUT: Ta = initial guess for equilibrium Hugoniot temperature (K) vb = desired equilibrium Hugoniot specific volume (m^3/kg) h1 = enthalpy at state 1 (J/kg) P1 = pressure at state 1 (Pa) v1 = specific volume at state 1 (m^3/kg) gas = working gas object

OUTPUT: enthalpy difference

SDtoolbox.Postshock.hug_fr — Method

hug_fr(x,vb,h1,P1,v1,gas)

Computes difference in enthalpy computed from assumed (T, V) state and fixed composition (frozen) hugoniot evaluation. Used with root solver such as 'fsolve' to compute frozen hugoniot as a function of volume. Input gas object is modified to correspond to input (T, V).

FUNCTION SYNTAX: diff = hug_fr(x,vb,h1,P1,v1,gas)

USAGE: fval = fsolve(hug_fr,Ta,args=(vb,h1,P1,v1,gas)) = frozen Hugoniot temperature (K) corresponding to vb

INPUT: Ta = initial guess for frozen Hugoniot temperature (K) vb = desired frozen Hugoniot specific volume (m^3/kg) h1 = enthalpy at state 1 (J/kg) P1 = pressure at state 1 (Pa) v1 = specific volume at state 1 (m^3/kg) gas = working gas object

OUTPUT: enthalpy difference

SDtoolbox.Postshock.partDeriv — Method

partDeriv(T,P,X,gas)

Calculates all needed variables related to partial derivatives, i.e γₛ (isentropic exponent), cₚ_eq (equilibrium specific heat at constant pressure), Dlvtp (d ln(v) / d ln(T) at constant pressure), Dlvpt (d ln(v) / d ln(P) at constant temperature). Derivatives are approximated via central differences.

SDtoolbox.Postshock.shk_calc — Method

shk_calc(U1, gas, gas1, ERRFT, ERRFV)

Calculates frozen post-shock state using Reynolds' iterative method.

FUNCTION SYNTAX: gas = shk_calc(U1,gas,gas1,ERRFT,ERRFV)

INPUT: U1 = shock speed (m/s) gas = working gas object gas1 = gas object at initial state ERRFT,ERRFV = error tolerances for iteration

OUTPUT: gas = gas object at frozen post-shock state

SDtoolbox.Postshock.shk_eq_calc — Method

shk_eq_calc(U1, gas, gas1, ERRFT, ERRFV)

Calculates equilibrium post-shock state using Reynolds' iterative method.

FUNCTION SYNTAX: gas = shk_calc(U1,gas,gas1,ERRFT,ERRFV)

INPUT: U1 = shock speed (m/s) gas = working gas object gas1 = gas object at initial state ERRFT,ERRFV = error tolerances for iteration

OUTPUT: gas = gas object at equilibrium post-shock state