ZND submodule

Documentation for SDtoolbox.jl

SDtoolbox.ZND — Module

Shock and Detonation Toolbox "ZND" module

Calculates ZND explosions.

This module defines the following functions:

znd!
zndsolve
getThermicity
soundspeed_fr

################################################################################ Theory, numerical methods and applications are described in the following report:

SDToolbox Numerical Tools for Shock and Detonation Wave Modeling,
Explosion Dynamics Laboratory, Contributors:
S. Browne, J. Ziegler, N. Bitter, B. Schmidt, J. Lawson and J. E. Shepherd,
GALCIT Technical Report FM2018.001 Revised January 2021.
California Institute of Technology, Pasadena, CA USA

Please cite this report and the website if you use these routines.

Please refer to LICENCE.txt or the above report for copyright and disclaimers.

http://shepherd.caltech.edu/EDL/PublicResources/sdt/

SDtoolbox.ZND.getThermicity — Method

getThermicity(gas::PyObject)

Returns the thermicity = sum ( (w/wi-hsi/(cpT))dyidt ).

FUNCTION SYNTAX: thermicity = getThermicity(gas)

INPUT: gas = Cantera gas object (not modified by this function)

OUTPUT: thermicity (1/s)

SDtoolbox.ZND.soundspeed_fr — Method

soundspeed_fr(gas::PyObject)

Computes the frozen sound speed by using a forward finite difference approximation and evaluating frozen composition states on the isentrope passing through the reference (S, V) state supplied by the gas object passed to the function.

FUNCTION SYNTAX: afrz = soundspeedfr(gas)

INPUT: gas = working gas object (restored to original state at end of function)

OUTPUT: afrz = frozen sound speed = sqrt({d P/d rho){s,x0})

SDtoolbox.ZND.znd! — Method

znd!(dy::Vector{Float64},y::Vector{Float64},params,t::Real)::Vector{Float64}

Set of ODEs to solve ZND Detonation Problem.

INPUT: t = time y = solution array [pressure, density, position, species mass 1, 2, ..] gas = working gas object U₁ = shock velocity (m/s) ρ₁ = initial density (kg/m^3)

OUTPUT: An array containing time derivatives of: pressure, density, distance and species mass fractions, formatted in a way that the integrator in zndsolve can recognize.

SDtoolbox.ZND.zndsolve — Method

zndsolve(gas::PyObject,gas₁::PyObject,U₁::Real; tend::Real=1e-3,maxstep::Real=1e-4,teval=nothing, relTol::Real=1e-5,absTol::Real=1e-8, advancedoutput::Bool=false,solver_algorithm=Rosenbrock23,)

ZND Model Detonation Computation Solves the set of ODEs defined in znd!().

FUNCTION SYNTAX: output = zndsolve(gas,gas1,U1,**kwargs)

INPUT gas = Cantera gas object - postshock state gas₁ = Cantera gas object - initial state U₁ = shock velocity (m/s)

OPTIONAL INPUT: tend = end time for integration, in sec maxstep = maximum time step for integration, in sec teval = array of time values to evaluate the solution at. If left as "None", solver will select values. Sometimes these may be too sparse for good-looking plots. relTol = relative tolerance absTol = absolute tolerance advancedoutput = calculates optional extra parameters such as induction lengths

OUTPUT: output = a dictionary containing the following results: time = time array distance = distance array

    T = temperature array
    P = pressure array
    rho = density array
    U = velocity array
    thermicity = thermicity array
    species = species mass fraction array

    M = Mach number array
    af = frozen sound speed array
    g = gamma (cp/cv) array
    wt = mean molecular weight array
    sonic = sonic parameter (c^2-U^2) array

    tfinal = final target integration time
    xfinal = final distance reached

    gas₁ = a copy of the input initial state
    U₁ = shock velocity

    and, if advanced_output=True:
    ind_time_ZND = time to maximum thermicity gradient
    ind_len_ZND = distance to maximum thermicity gradient
    exo_time_ZND = pulse width (in secs) of thermicity  (using 1/2 max)
    ind_time_ZND = pulse width (in meters) of thermicity (using 1/2 max)
    max_thermicity_width_ZND = according to Ng et al definition